requirements.txt
wheel==0.33.1
google-ads==1.1.0
pandas==0.24.2
Flask==1.0.2